[FEAT/#378] 코드리뷰 반영 (firebase 모듈 사용)
