+++ core/firebase/src/main/java/com/terning/core/firebase/fcmtoken/FcmTokenProvider.kt
@@ -1,0 +1,5 @@
+package com.terning.core.firebase.fcmtoken
+
+interface FcmTokenProvider {
+    suspend fun fetchFcmToken(): String
+}
