+++ data/token/src/main/java/com/terning/data/token/repositoryimpl/TokenRepositoryImpl.kt
@@ -1,30 +1,33 @@
 package com.terning.data.token.repositoryimpl
 
+import com.terning.core.firebase.fcmtoken.FcmTokenProvider
 import com.terning.core.local.TerningDataStore
 import com.terning.domain.token.repository.TokenRepository
 import javax.inject.Inject
 
 class TokenRepositoryImpl @Inject constructor(
-    private val terningDataStore: TerningDataStore
+    private val terningDataStore: TerningDataStore,
+    private val fcmTokenProvider: FcmTokenProvider
 ) : TokenRepository {
     override fun getAccessToken(): String = terningDataStore.accessToken
-
-    override fun getRefreshToken(): String = terningDataStore.refreshToken
-    override fun getFcmToken(): String = terningDataStore.fcmToken
-
-    override fun getUserId(): Long = terningDataStore.userId
 
     override fun setAccessToken(accessToken: String) {
         terningDataStore.accessToken = accessToken
     }
 
+    override fun getRefreshToken(): String = terningDataStore.refreshToken
+
     override fun setRefreshToken(refreshToken: String) {
         terningDataStore.refreshToken = refreshToken
     }
 
-    override fun setFcmToken(fcmToken: String) {
-        terningDataStore.fcmToken = fcmToken
+    override fun getFcmToken(): String = terningDataStore.fcmToken
+
+    override suspend fun fetchAndSetFcmToken(): Result<Unit> = runCatching {
+        terningDataStore.fcmToken = fcmTokenProvider.fetchFcmToken()
     }
+
+    override fun getUserId(): Long = terningDataStore.userId
 
     override fun setUserId(userId: Long) {
         terningDataStore.userId = userId
