+++ domain/token/src/main/java/com/terning/domain/token/repository/TokenRepository.kt
@@ -3,19 +3,19 @@ package com.terning.domain.token.repository
 interface TokenRepository {
     fun getAccessToken(): String
 
-    fun getRefreshToken(): String
-
-    fun getFcmToken(): String
-
     fun setAccessToken(accessToken: String)
+
+    fun getRefreshToken(): String
 
     fun setRefreshToken(refreshToken: String)
 
-    fun setFcmToken(fcmToken: String)
+    fun getFcmToken(): String
 
-    fun setUserId(userId: Long)
+    suspend fun fetchAndSetFcmToken(): Result<Unit>
 
     fun getUserId(): Long
+
+    fun setUserId(userId: Long)
 
     fun clearInfo()
 }
